requirements.txt
# Core OCR and Image Processing
paddleocr>=2.7.0
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
pdf2image>=1.16.0

# Deep Learning and Transformers
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0

# Table Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Mathematical OCR
# texteller  # Will need to be installed separately or from source

# OpenAI for diagrams only
openai>=1.3.0

# Document Processing
PyMuPDF>=1.23.0  # Alternative to pdf2image
python-docx>=0.8.11

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Optional: For better performance
onnxruntime-gpu>=1.16.0  # If using GPU